+++ ecommerce/frontend/src/components/layout/Content.js
@@ -7,7 +7,7 @@ import { useAppContext } from '../../context/AppContext'
 import { CategoryBar } from '../CategoryBar'
 import { EmailVerification } from '../EmailVerification'
 import { IsVerifiedNotification } from '../IsVerifiedNotification'
-import { OrdersList } from '../OrdersList'
+import { OrdersList } from '../order/OrdersList'
 import { CheckoutPage } from '../pages/CheckoutPage'
 import { HomePage } from '../pages/HomePage'
 import { LoginPage } from '../pages/LoginPage'
